Reduce `SectorIndex` from `u64` to `u16`, which will be still sufficient for 64TiB plot, for larger amount of space more plots will be needed

+++ crates/subspace-farmer/src/single_disk_plot/plotting.rs
@@ -82,7 +82,10 @@ where
         };
         let mut sector_metadata = unsafe {
             MmapOptions::new()
-                .offset(RESERVED_PLOT_METADATA + (sector_index * sector_metadata_size as u64))
+                .offset(
+                    RESERVED_PLOT_METADATA
+                        + (u64::from(sector_index) * sector_metadata_size as u64),
+                )
                 .len(sector_metadata_size)
                 .map_mut(&metadata_file)?
         };
